+++ CHANGELOG.md
@@ -1,3 +1,77 @@
+# CHANGE LOG (3.19.0) - 2.11.5
+
+## Application Client
+
+
+
+### Order
+
+
+
+#### getRefundModesWithPriceBreakup
+
+- ##### What's New
+	- [Added] method <code>getRefundModesWithPriceBreakup</code>
+
+
+
+### Payment
+
+
+
+#### getPosPaymentModeRoutes
+
+- ##### What's New
+	- [Added] <code>query</code> parameter <code>fulfillmentOption</code> (type: <code>array</code>)
+
+
+
+## Platform Client
+
+
+
+### Order
+
+
+
+#### addStateManagerConfig
+
+- ##### What's New
+	- [Added] property <code>data.configs[].preHooks[].filter</code> of schema <code>PreHook</code> in request body
+	- [Added] property <code>data.configs[].postHooks[].filter</code> of schema <code>PostHook</code> in request body
+
+- ##### What's Deprecated
+	- [Breaking] [Deleted] property <code>data.configs[].preHooks[].filters</code> of schema <code>PreHook</code> in request body
+	- [Breaking] [Deleted] property <code>data.configs[].postHooks[].filters</code> of schema <code>PostHook</code> in request body
+
+
+#### getStateManagerConfig
+
+- ##### What's New
+	- [Added] property <code>data[].configs[].preHooks[].filter</code> of schema <code>PreHook</code> in response with status code 200
+	- [Added] property <code>data[].configs[].postHooks[].filter</code> of schema <code>PostHook</code> in response with status code 200
+
+- ##### What's Deprecated
+	- [Breaking] [Deleted] property <code>data[].configs[].preHooks[].filters</code> of schema <code>PreHook</code> in response with status code 200
+	- [Breaking] [Deleted] property <code>data[].configs[].postHooks[].filters</code> of schema <code>PostHook</code> in response with status code 200
+
+
+### User
+
+
+
+#### createUser
+
+- ##### What's New
+	- [Added] property <code>dob</code> of schema <code>CreateUserRequestSchema</code> in request body
+
+
+#### updateUser
+
+- ##### What's New
+	- [Added] property <code>dob</code> of schema <code>UpdateUserRequestSchema</code> in request body
+
+
 # CHANGE LOG (3.18.0) - 2.11.3
 
 ## Application Client